+++ src/LiveComponent/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Add support for [Symfony UID](https://symfony.com/doc/current/components/uid.html) hydration/dehydration
 -   `ComponentWithFormTrait` now correctly checks for a `TranslatableInterface` placeholder for `<select>` elements
+-   Fix `LiveComponentHydrator::hydrateValue()` to hydrate null values
 -   Add property name as second paramter to LiveProp modifier callback
 
 ## 2.23.0